`launch` v2: fly pg and upstash are no longer top-level plan items
also split postgres and redis plans into new files
