+++ internal/command/launch/plan.go
@@ -1,15 +1,10 @@
 package launch
 
 import (
-	"context"
-	"fmt"
-
-	"github.com/samber/lo"
 	"github.com/superfly/flyctl/api"
-	"github.com/superfly/flyctl/client"
-	"github.com/superfly/flyctl/gql"
-	"github.com/superfly/flyctl/terminal"
 )
+
+const descriptionNone = "<none>"
 
 type launchPlanSource struct {
 	appNameSource  string
@@ -32,9 +27,9 @@ type launchPlan struct {
 	MemoryMB int    `json:"vm_memory,omitempty" url:"vm_memory,omitempty"`
 	VmSize   string `json:"vm_size,omitempty" url:"vm_size,omitempty"`
 
-	Postgres *postgresPlan `json:"-"` // `json:"postgres" url:"postgres"`
+	Postgres postgresPlan `json:"postgres" url:"postgres"`
 
-	Redis *redisPlan `json:"-"` // `json:"redis" url:"redis"`
+	Redis redisPlan `json:"redis" url:"redis"`
 
 	ScannerFamily string `json:"scanner_family" url:"scanner_family"`
 }
@@ -57,54 +52,4 @@ func (p *launchPlan) SetGuestFields(guest *api.MachineGuest) {
 	p.CPUKind = guest.CPUKind
 	p.MemoryMB = guest.MemoryMB
 	p.VmSize = guest.ToSize()
-}
-
-// TODO
-type postgresPlan struct {
-	VmSize     string `json:"vm_size" url:"vm_size"`
-	Nodes      int    `json:"nodes" url:"nodes"`
-	DiskSizeGB int    `json:"disk_size_gb" url:"disk_size_gb"`
-}
-
-func (p *postgresPlan) Guest() *api.MachineGuest {
-	guest := api.MachineGuest{}
-	guest.SetSize(p.VmSize)
-	return &guest
-}
-
-func (p *postgresPlan) Describe() string {
-	if p == nil {
-		return "<none>"
-	}
-	nodePlural := lo.Ternary(p.Nodes == 1, "", "s")
-	return fmt.Sprintf("%d Node%s, %s, %dGB disk", p.Nodes, nodePlural, p.Guest().String(), p.DiskSizeGB)
-}
-
-type redisPlan struct {
-	PlanId       string   `json:"plan_id" url:"plan_id"`
-	Eviction     bool     `json:"eviction" url:"eviction"`
-	ReadReplicas []string `json:"read_replicas" url:"read_replicas"`
-}
-
-func (p *redisPlan) Describe(ctx context.Context) (string, error) {
-	if p == nil {
-		return "<none>", nil
-	}
-
-	client := client.FromContext(ctx)
-
-	result, err := gql.ListAddOnPlans(ctx, client.API().GenqClient)
-	if err != nil {
-		terminal.Debugf("Failed to list addon plans: %s\n", err)
-		return "<internal error>", err
-	}
-
-	for _, plan := range result.AddOnPlans.Nodes {
-		if plan.Id == p.PlanId {
-			evictionStatus := lo.Ternary(p.Eviction, "enabled", "disabled")
-			return fmt.Sprintf("%s: %s Max Data Size, ($%d / month), eviction %s", plan.DisplayName, plan.MaxDataSize, plan.PricePerMonth, evictionStatus), nil
-		}
-	}
-
-	return "<plan not found, this is an error>", fmt.Errorf("plan not found: %s", p.PlanId)
 }
